publishing partial untested progress
